29882 is most easily handled by a hand-modded fi file - set up for that.
svn:revision:72658
svn:branch:trunk
svn:account:starseeker

+++ misc/repoconv/svnfexport/common_structs.h
@@ -127,13 +127,16 @@ int verify_repos(long int rev, std::string branch_svn, std::string branch_git)
     std::system(svn_emptydir_rm.c_str());
     std::system(git_setup.c_str());
     if (rev > starting_rev) {
+	struct stat buffer;
 	std::string fi_file = std::to_string(rev) + std::string(".fi");
-	std::string git_fi = std::string("cd cvs_git_working && cat ../") + fi_file + std::string(" | git fast-import && git reset --hard HEAD && cd ..");
-	if (std::system(git_fi.c_str())) {
-	    std::string failed_file = std::string("failed-") + fi_file;
-	    std::cout << "Fatal - could not apply fi file for revision " << rev << "\n";
-	    rename(fi_file.c_str(), failed_file.c_str());
-	    exit(1);
+	if (stat(fi_file.c_str(), &buffer) == 0) {
+	    std::string git_fi = std::string("cd cvs_git_working && cat ../") + fi_file + std::string(" | git fast-import && git reset --hard HEAD && cd ..");
+	    if (std::system(git_fi.c_str())) {
+		std::string failed_file = std::string("failed-") + fi_file;
+		std::cout << "Fatal - could not apply fi file for revision " << rev << "\n";
+		rename(fi_file.c_str(), failed_file.c_str());
+		exit(1);
+	    }
 	}
     }
     std::system(git_clone.c_str());
